+++ src/converters/piral-million/convert.ts
@@ -1,0 +1,29 @@
+import { createConverter } from './esm/converter';
+
+export interface HtmlComponent<TProps> {
+  component: {
+    mount(element: HTMLElement, props: TProps, ctx: any, locals: any): void;
+    update?(element: HTMLElement, props: TProps, ctx: any, locals: any): void;
+    unmount?(element: HTMLElement, locals: any): void;
+  };
+  type: 'html';
+}
+
+export interface MillionConverter {
+  (...params: Parameters<ReturnType<typeof createConverter>>): HtmlComponent<any>;
+}
+
+export function createMillionConverter(...params: Parameters<typeof createConverter>) {
+  const convert = createConverter(...params);
+  const Extension = convert.Extension;
+  const from: MillionConverter = (root) => ({
+    type: 'html',
+    component: convert(root),
+  });
+
+  return { from, Extension };
+}
+
+const { from: fromMillion, Extension: MillionExtension } = createMillionConverter();
+
+export { fromMillion, MillionExtension };